UVITTools.jl

Tools for AstroSat/UVIT grating spectroscopy and photometry.

Package Features

Extract flux calbrated UVIT FUV/NUV grating spectra,

generate source and background PHA spectral data compatible with XSPEC/Sherpa

Perform aperture photometry of sources in the UVIT images in any of the broadband filters

Saturation for point sources

Function Documentation

FUV-Grating1

UVITTools.fuv_grating1_count_spec — Function

fuv_grating1_count_spec(fuv_grating1_image_file, ds9regfile[, order = -1, angle_xaxis_disp_deg=0.0, cross_disp_width_pixels = 40, rate = true])

Extract count rate spectrum from AstroSat/UVIT FUV-Grating1 dispersed image generated from CCDLAB processing pipeline.

...

Arguments

Required parameters

fuv_grating1_image_file::String: Name of the FUV-Grating1 image file in FITS format generated using CCDLAB.

ds9regfile::String: Name of the ds9 region file with center as the zero order position.

Optional parameters

order::Int: -2 (default), Grating order to be used to extract the spectrum. Allowed orders=-1 and -2.

angle_xaxis_disp_deg: 0.0 (default), angle between dispersion axis and x-axis.

cross_disp_width_pixels::String: 40 (default), width in pixels in the cross-dispersion direction.

rate::Bool: true (default) for count rate spectrum, otherwise false for count spectrum.

Output

count spectrum file

region file compatible with ds9 that can be used to verify the extraction region for requested order.

Count spectrum as (pixel numbers relative to zero order, counts/s or counts, errors)

...

UVITTools.fuv_grating1_net_countrate_spec — Function

fuv_grating1_net_countrate_spec(fuv_grating1_image_file::String, ds9srcregfile::String, ds9bgdregfile::String; order::Int = -2, angle_xaxis_disp_deg::Float64=0.0, cross_disp_width_pixels::Int = 40)

fuv_grating1_net_countrate_spec(fuv_grating1_image_file, ds9srcregfile, ds9bgdregfile[, order = -2, angle_xaxis_disp_deg=0.0, cross_disp_width_pixels = 40])

Extract background corrected, net count rate spectrum from AstroSat/UVIT FUV-Grating1 dispersed image generated from CCDLAB processing pipeline.

...

Arguments

Required parameters

fuv_grating1_image_file::String: Name of the FUV-Grating1 image file in FITS format generated using CCDLAB.

ds9srcregfile::String: Name of the ds9 region file with source center as the zero order position.

ds9bgdregfile::String: Name of the ds9 region file with center in a source-free region of the image.

Optional parameters

order::Int: -2 (default), Grating order to be used to extract the spectrum. Allowed orders=-1 and -2.

angle_xaxis_disp_deg: 0.0 (default), angle between dispersion axis and x-axis.

cross_disp_width_pixels::Int: 40 (default), width in pixels in the cross-dispersion direction.

Output

DS9 Region files used for extraction source and background count spectra.

Source and background count spectra in ascii files.

...

UVITTools.fuv_grating1_pixel2lamA — Function

fuv_grating1_pixel2lamA(pixel_num_wrt_zero_order[, order = -2])

Convert pixel number relative to zero to wavelength in Angstrom.

This function is used for wavelength calibration of FUV-Grating1 count spectrum.

Required parameters

pixel_num_wrt_zero_order::Int: Pixel numbers relative to zero order.

Optional parameters

order::Int: -2 (default), Grating order. Allowed orders=-1 and -2.

...

UVITTools.fuv_grating1_wavelength_calib — Function

fuv_grating1_wavelength_calib(pixels, netsrc_spec_counts_per_s[, order = -2])

Convert pixel numbers relative to zero order to wavelengths.

This function is used for wavelength calibration of FUV-Grating1 count spectrum.

...

Arguments

Required parameters

pixels::Array: An array of pixel numbers relative to zero order.

netsrc_spec_counts_per_s::Array: An array of net count rates corresponding to the relative pixel numbers.

Optional parameters

order::Int: -2 (default), Grating order. Allowed orders=-1 and -2.

...

UVITTools.fuv_grating1_flux_calib — Function

fuv_grating1_flux_calib(lamA, netsrc_spec_counts_per_s_A[,order=-2])

Flux calibrate the wavelength-calibrated count spectrum from FUV-Grating1.

This function calculates the effective area at each wavelength of the count spectrum, then uses the effective areas to convert net count rates to f_λ in CGS units.

...

Arguments

Required

-lamA::Array{Float64}: Array of wavelengths in Å. -netsrc_spec_counts_per_s_A::Array{Float64}: Array of background corrected counts/s/Å corresponding to wavelength array.

Optional

-Order::Number: Grating order -2 (default) or -1.

...

UVITTools.fuv_grating1_fluxed_spec — Function

fuv_grating1_fluxed_spec(target,fuv_grating1_image_file, ds9srcregfile, ds9bgdregfile[, order = -2, angle_xaxis_disp_deg=0.0, cross_disp_width_pixels = 40])

Extract flux calibrated spectrum from AstroSat/UVIT FUV-Grating1 dispersed image generated from CCDLAB processing pipeline.

This is a main function that uses other functions for extraction of source and background spectra, wavelenth and flux calibrations, and outputs fluxed spectrum. For details on grating orders, wavelength and flux calibrations, see Dewangan (2021). ...

Arguments

Required parameters

target::String: Name of the target available in the observed UVIT field.

fuv_grating1_image_file::String: Name of the FUV-Grating1 image file in FITS format generated using CCDLAB.

ds9srcregfile::String: Name of the ds9 region file with source center as the zero order position.

ds9bgdregfile::String: Name of the ds9 region file with center in a source-free region of the image.

Optional parameters

order::Int: -2 (default), Grating order to be used to extract the spectrum. Allowed orders=-1 and -2.

angle_xaxis_disp_deg::Float64: 0.0 (default), angle between dispersion axis and x-axis.

cross_disp_width_pixels::String: 40 (default), width in pixels in the cross-dispersion direction.

Output

(λ, f_λ, err_f_λ)

Fluxed spectrum saved in an ascii file.

DS9 source and background region files used for spectral extraction.

Source and background count spectral data in ascii files

Net source count spectrum in ascii file

...

UVITTools.fuv_grating1_phafile — Function

fuv_grating1_phafile(target,fuv_grating1_image_file, ds9srcregfile, ds9bgdregfile[,order = -2, angle_xaxis_disp_deg=0.0, cross_disp_width_pixels= 40])

Extract XSPEC/Sherpa compatible source and background PHA spectral files from AstroSat/UVIT FUV-Grating1 dispersed image generated from CCDLAB processing pipeline.

This function extracts source and background count spectra using the zero order positions provided in the DS9 region files and converts them into PHA spectral files. For details on grating orders and spectral responses, see Dewangan (2021).

...

Arguments

Required parameters

target::String: Name of the target available in the observed UVIT field.

fuv_grating1_image_file::String: Name of the FUV-Grating1 image file in FITS format generated using CCDLAB.

ds9srcregfile::String: Name of the ds9 region file with source center as the zero order position.

ds9bgdregfile::String: Name of the ds9 region file with center in a source-free region of the image.

Optional parameters

order::Int: -2 (default), Grating order to be used to extract the spectrum. Allowed orders=-1 and -2.

angle_xaxis_disp_deg::Float64: 0.0 (default), angle between dispersion axis and x-axis.

cross_disp_width_pixels::Int: 40 (default), width in pixels in the cross-dispersion direction.

#read necessary keywords`

Output

Source and background PHA files with RESPFILE key updated in the source spectral file

Source and background DS9 region files used extract count spectra

Source and Background count spectra in ascii files

Net source count spectrum in ascii file

Some relevant information are also printed on the screen.

...

UVITTools.fuv_grating1_ea — Function

fuv_grating1_ea(lamA[,order = -2])

Calculate effective area in cm^2 at a desired wavelength (Angstrom) and grating order -1 or -2.

The calculation of the effective area is based on the updated grating calibration using the procedures described in Dewangan (2021). This function is used for flux calibration of count spectrum.

...

Arguments

Required

lam::Number: Wavelength in Angstrom.

Optional

-order::Int: Grating order -1 or -2. ...

Example

julia> fuv_grating1_ea(1450.4,order=-2)
4.071442845470301

FUV-Grating2

UVITTools.fuv_grating2_count_spec — Function

fuv_grating2_count_spec(fuv_grating2_image_file, ds9regfile[, order = -2, disp_aligned_to_xaxis = false, angle_xaxis_disp_deg = 267.479, cross_disp_width_pixels = 40, rate = true])

Extract count rate spectrum from AstroSat/UVIT FUV-Grating2 dispersed image generated from CCDLAB processing pipeline.

...

Arguments

Required parameters

fuv_grating2_image_file::String: Name of the FUV-Grating2 image file in FITS format generated using CCDLAB.

ds9regfile::String: Name of the ds9 region file with center as the zero order position.

Optional parameters

order::Int: -2 (default), Grating order to be used to extract the spectrum. Allowed orders=-1 and -2.

disp_aligned_to_xaxis: false (default), true if the dispersion axis is rotated to align to x-axis in the detector coordinates, false if dispersion axis left unrotated.

angle_xaxis_disp_deg: 267.479 (default), the angle in degrees between dispersion direction (from zero to minus orders) and the x-axis.

If the dispersion axis is left unrotated in the detector coordiantes, the default value should be appropriate. 
	If the dispersion axis is rotated to align to x-axis, the angle should be set to 180 or zero degrees.

cross_disp_width_pixels::Int: 40 (default), width in pixels in the cross-dispersion direction.

rate::Bool: true (default) for count rate spectrum, otherwise false for count spectrum.

Output

count spectrum file

region file compatible with ds9 that can be used to verify the extraction region.

Count spectrum as (pixel numbers relative to zero order, counts/s or counts, errors)

...

UVITTools.fuv_grating2_net_countrate_spec — Function

fuv_grating2_net_countrate_spec(fuv_grating2_image_file, ds9srcregfile, ds9bgdregfile[, order = -2, disp_aligned_to_xaxis=false, angle_xaxis_disp_deg=267.479, cross_disp_width_pixels = 40])

Extract background corrected, net count rate spectrum from AstroSat/UVIT FUV-Grating2 dispersed image generated from CCDLAB processing pipeline.

...

Arguments

Required parameters

fuv_grating2_image_file::String: Name of the FUV-Grating2 image file in FITS format generated using CCDLAB.

ds9srcregfile::String: Name of the ds9 region file with source center as the zero order position.

ds9bgdregfile::String: Name of the ds9 region file with center in a source-free region of the image.

Optional parameters

order::Int: -2 (default), Grating order to be used to extract the spectrum. Allowed orders=-1 and -2.

disp_aligned_to_xaxis: false (default), true if the dispersion axis is rotated to align to x-axis in the detector coordinates, false if dispersion axis left unrotated.

angle_xaxis_disp_deg: 267.479 (default), the angle in degrees between dispersion direction (from zero to minus orders) and the x-axis.

If the dispersion axis is left unrotated in the detector coordiantes, the default value should be appropriate. 
	If the dispersion axis is rotated to align to x-axis, the angle should be set to 180 or zero degrees.

cross_disp_width_pixels::Int: 40 (default), width in pixels in the cross-dispersion direction.

...

UVITTools.fuv_grating2_pixel2lamA — Function

fuv_grating2_pixel2lamA(pixel_num_wrt_zero_order[, order = -2])

Convert FUV-Grating2 pixel number relative to zero order to wavelength in Angstrom.

This function is used for wavelength calibration of FUV-Grating2 count spectrum.

Required parameters

pixel_num_wrt_zero_order::Int: Pixel numbers relative to zero order.

Optional parameters

order::Int: -2 (default), Grating order. Allowed orders=-1 and -2.

...

UVITTools.fuv_grating2_wavelength_calib — Function

fuv_grating2_wavelength_calib(pixels, netsrc_spec_counts_per_s[, order = -2])

Convert pixel numbers relative to zero to wavelengths.

This function is used for wavelength calibration of FUV-Grating2 count spectrum.

...

Arguments

Required parameters

pixels::Array: An array of pixel numbers relative to zero order.

netsrc_spec_counts_per_s::Array: An array of net count rates corresponding to the relative pixel numbers.

Optional parameters

order::Int: -2 (default), Grating order. Allowed orders=-1 and -2.

...

UVITTools.fuv_grating2_flux_calib — Function

fuv_grating2_flux_calib(lamA, netsrc_spec_counts_per_s_A[,order=-2])

Flux calibrate the wavelength-calibrated count spectrum from FUV-Grating2.

This function calculates the effective area at each wavelength of the count spectrum, then uses the effective areas to convert net count rates to f_λ in CGS units.

...

Arguments

Required

-lamA::Array{Float64}: Array of wavelengths in Å. -netsrc_spec_counts_per_s_A::Array{Float64}: Array of background corrected counts/s/Å corresponding to wavelength array.

Optional

-Order::Number: Grating order -2 (default) or -1. ...

UVITTools.fuv_grating2_fluxed_spec — Function

fuv_grating2_fluxed_spec(target,fuv_grating2_image_file, ds9srcregfile, ds9bgdregfile[,order = -2, disp_aligned_to_xaxis=false, angle_xaxis_disp_deg=267.479, cross_disp_width_pixels= 40])

Extract flux calibrated spectrum from AstroSat/UVIT FUV-Grating2 dispersed image generated from CCDLAB processing pipeline.

This is a main function that uses other functions for extraction of source and background spectra, wavelenth and flux calibrations, and outputs fluxed spectrum. For details on grating orders, wavelength and flux calibrations, see Dewangan (2021). ...

Arguments

Required parameters

target::String: Name of the target available in the observed UVIT field.

fuv_grating2_image_file::String: Name of the FUV-Grating2 image file in FITS format generated using CCDLAB.

ds9srcregfile::String: Name of the ds9 region file with source center as the zero order position.

ds9bgdregfile::String: Name of the ds9 region file with center in a source-free region of the image.

Optional parameters

order::Int: -2 (default), Grating order to be used to extract the spectrum. Allowed orders=-1 and -2.

disp_aligned_to_xaxis: false (default), true if the dispersion axis is rotated to align to x-axis in the detector coordinates, false if dispersion axis left unrotated.

angle_xaxis_disp_deg: 267.479 (default), the angle in degrees between dispersion direction (from zero to minus orders) and the x-axis.

If the dispersion axis is left unrotated in the detector coordiantes, the default value should be appropriate. 
	If the dispersion axis is rotated to align to x-axis, the angle should be set to 180 or zero degrees.

cross_disp_width_pixels::String: 40 (default), width in pixels in the cross-dispersion direction.

Output

(λ, f_λ, err_f_λ)

Fluxed spectrum saved in an ascii file.

...

UVITTools.fuv_grating2_phafile — Function

fuv_grating2_phafile(target,fuv_grating2_image_file, ds9srcregfile, ds9bgdregfile[, order = -2, disp_aligned_to_xaxis=false, angle_xaxis_disp_deg=267.479, cross_disp_width_pixels= 40])

Extract XSPEC/Sherpa compatible source and background PHA spectral files from AstroSat/UVIT FUV-Grating2 dispersed image generated from CCDLAB processing pipeline.

This function extracts source and background count spectra using the zero order positions provided in the DS9 region files and converts them into PHA spectral files. For details on grating orders and spectral responses, see Dewangan (2021).

...

Arguments

Required parameters

target::String: Name of the target available in the observed UVIT field.

fuv_grating2_image_file::String: Name of the FUV-Grating1 image file in FITS format generated using CCDLAB.

ds9srcregfile::String: Name of the ds9 region file with source center as the zero order position.

ds9bgdregfile::String: Name of the ds9 region file with center in a source-free region of the image.

Optional parameters

order::Int: -2 (default), Grating order to be used to extract the spectrum. Allowed orders=-1 and -2.

disp_aligned_to_xaxis: false (default), true if the dispersion axis is rotated to align to x-axis in the detector coordinates, false if dispersion axis left unrotated.

angle_xaxis_disp_deg: 267.479 (default), the angle in degrees between dispersion direction (from zero to minus orders) and the x-axis.

If the dispersion axis is left unrotated in the detector coordiantes, the default value should be appropriate. 
	If the dispersion axis is rotated to align to x-axis, the angle should be set to 180 or zero degrees.

cross_disp_width_pixels::Int: 40 (default), width in pixels in the cross-dispersion direction.

Output

Source and background PHA files.

Some relevant information are also printed on the screen.

...

UVITTools.fuv_grating2_ea — Function

fuv_grating2_ea(lamA[,order = -2])

Calculate FUV-Grating2 effective area in cm^2 at a desired wavelength (Angstrom) and grating order -1 or -2.

The calculation of the effective area is based on the grating calibration peformed in Dewangan (2021). This function is used for flux calibration of count spectrum.

...

Arguments

Required

lam::Number: Wavelength in Angstrom.

Optional

-order::Int: Grating order -1 or -2. ...

Example

julia> fuv_grating2_ea(1450.4,order=-2)
4.076758827954109

NUV-Grating

UVITTools.nuv_grating_m1_count_spec — Function

nuv_grating_m1_count_spec(nuv_grating_image_file, ds9regfile[, cross_disp_width_pixels = 40, rate = true, outfile="nuv_grating_m1_count_spec.dat"])

Extract count rate spectrum from AstroSat/UVIT NUV-Grating dispersed image generated from CCDLAB processing pipeline.

...

Arguments

Required parameters

nuv_grating_image_file::String: Name of the NUV-Grating image file in FITS format generated using CCDLAB.

ds9regfile::String: Name of the ds9 region file with center as the zero order position.

Optional parameters

order::Int: -2 (default), Grating order to be used to extract the spectrum. Allowed orders=-1 and -2.

cross_disp_width_pixels::String: 40 (default), width in pixels in the cross-dispersion direction.

rate::Bool: true (default) for count rate spectrum, otherwise false for count spectrum.

outfile::String: Name of ascii output file name. Default file name: "nuv_grating_m1_count_spec.dat".

...

UVITTools.nuv_grating_m1_net_countrate_spec — Function

nuv_grating_m1_net_countrate_spec(nuv_grating_image_file, ds9srcregfile, ds9bgdregfile[, cross_disp_width_pixels = 40, outfile="nuv_grating_m1_net_countrate_spec.dat"])

Extract background corrected, net count rate spectrum from AstroSat/UVIT NUV-Grating dispersed image generated from CCDLAB processing pipeline.

...

Arguments

Required parameters

nuv_grating_image_file::String: Name of the NUV-Grating image file in FITS format generated using CCDLAB.

ds9srcregfile::String: Name of the ds9 region file with source center as the zero order position.

ds9bgdregfile::String: Name of the ds9 region file with center in a source-free region of the image.

Optional parameters

cross_disp_width_pixels::String: 40 (default), width in pixels in the cross-dispersion direction.

outfile::String: Name of ascii output file name. Default file name: "nuv_grating_m1_net_countrate_spec.dat".

...

UVITTools.nuv_grating_m1_wavelength_calib — Function

nuv_grating_m1_wavelength_calib(pixels, netsrc_spec_counts_per_s)

Convert pixel numbers relative to zero to wavelengths.

This function is used for wavelength calibration of NUV-Grating order=-1 count spectrum.

...

Arguments

Required parameters

pixels::Array: An array of pixel numbers relative to zero order.

netsrc_spec_counts_per_s::Array: An array of net count rates corresponding to the relative pixel numbers.

Optional parameters

None

...

UVITTools.nuv_grating_m1_flux_calib — Function

nuv_grating_m1_flux_calib(lamA, netsrc_spec_counts_per_s_A)

Flux calibrate the wavelength-calibrated count spectrum from NUV-Grating1 order=-1.

This function calculates the effective area at each wavelength of the count spectrum, then uses the effective areas to convert net count rates to f_λ in CGS units.

...

Arguments

Required

-lamA::Array{Float64}: Array of wavelengths in Å. -netsrc_spec_counts_per_s_A::Array{Float64}: Array of background corrected counts/s/Å corresponding to wavelength array.

Optional

None.

...

UVITTools.nuv_grating_m1_fluxed_spec — Function

nuv_grating_m1_fluxed_spec(target,nuv_grating_image_file, ds9srcregfile, ds9bgdregfile[,order=-1, cross_disp_width_pixels= 40, outfile="default"])

Extract flux calibrated spectrum from AstroSat/UVIT NUV-Grating order=-1 dispersed image generated from CCDLAB processing pipeline.

This is a main function that uses other functions for extraction of source and background spectra, wavelenth and flux calibrations, and outputs fluxed spectrum. For details on grating orders, wavelength and flux calibrations, see Dewangan (2021). ...

Arguments

Required parameters

target::String: Name of the target available in the observed UVIT field.

nuv_grating_image_file::String: Name of the NUV-Grating image file in FITS format generated using CCDLAB.

ds9srcregfile::String: Name of the ds9 region file with source center as the zero order position.

ds9bgdregfile::String: Name of the ds9 region file with center in a source-free region of the image.

Optional parameters

order::Int: -1 (default), Grating order to be used to extract the spectrum. Allowed orders=-1. Order=-2 is not yet calibrated.

cross_disp_width_pixels::String: 40 (default), width in pixels in the cross-dispersion direction.

-outfile::String:Name of the output file. A default file name based on the target name/grating will be generated if outfile="default".

Output

(λ, f_λ, err_f_λ)

Fluxed spectrum saved in an ascii file.

...

UVITTools.nuv_grating_phafile — Function

nuv_grating_phafile(target,nuv_grating1_image_file, ds9srcregfile, ds9bgdregfile[, cross_disp_width_pixels= 40])

Extract XSPEC/Sherpa compatible source and background PHA spectral files from AstroSat/UVIT NUV-Grating dispersed image generated from CCDLAB processing pipeline.

This function extracts source and background count spectra using the zero order positions provided in the DS9 region files and converts them into PHA spectral files. For details on grating orders and spectral responses, see Dewangan (2021).

...

Arguments

Required parameters

target::String: Name of the target available in the observed UVIT field.

nuv_grating_image_file::String: Name of the NUV-Grating image file in FITS format generated using CCDLAB.

ds9srcregfile::String: Name of the ds9 region file with source center as the zero order position.

ds9bgdregfile::String: Name of the ds9 region file with center in a source-free region of the image.

Optional parameters

cross_disp_width_pixels::String: 40 (default), width in pixels in the cross-dispersion direction.

respdir::String: Name of the directory containing the response matrices in the local machine. The response files can be downloaded from the GitHub page.

#read necessary keywords`

Output

Source and background PHA files.

Some relevant information are also printed on the screen.

...

Aperture Photometry

UVITTools.read_ds9reg — Function

read_ds9reg(ds9regfile)

Read a DS9 region file.

This function reads a region file created with the DS9 tool, and outputs the region parameters. ...

Arguments

Required parameters

ds9regfile::String : DS9 region file for a circular, annular or elliptical region.

...

UVITTools.uvit_aphot — Function

uvit_aphot(imagefile,ds9srcregion, ds9bgdregion[, satu_corr=true, mst_or_bjd="mst"])

Perform aperture photometry of a source present in an UVIT FUV/NUV FITS image obtained from CCDLAB pipeline.

This function uses DS9 source and background region files and extracts count rates, corrects for any saturation, calculates the background net count rate, and then converts to flux density f_λ and AB magnitude using calibration information available in Tandon et al. (2020).

If the source region is a circle and satucorr is true, then the source count rate is also corrected for saturation using the function `uvitsaturation_corr.jl` (see below). For details, see Tandon et al. (2017).

...

Arguments

Required parameters

imagefile::String: Name of the FUV or NUV image file in FITS format generated using CCDLAB.

ds9srcregfile::String: Name of the ds9 region file for the source.

ds9bgdregfile::String: Name of the ds9 region file for background.

Optional parameters

satu_corr::Bool: true (default), Apply saturation correction only if source extraction region is circular. Valid for point sources only and if the redius of the circular extraction region is about 29 pixels or 12 arcsec (see Tandon et al. 2020).

mst_or_bjd::String: Mission time or barycentric julian day, default: "mst".

...

UVITTools.uvit_countrate2flux — Function

uvit_countrate2flux(channel, filter, cps)

Convert net source count rate to flux density or magnitude for any of FUV/NUV broadband filters. This function is used by uvit_aphot.jl.

...

Arguments

channel::String: UVIT channel FUV or NUV

filter::String : FUV or NUV filter e.g., "BaF2".

cps::Float64 : Counts per second.

...

UVITTools.uvit_filter2pha — Function

uvit_filter2pha(target,imagefile, ds9srcregfile, ds9bgdregfile[,satu_corr=true, respdir=""/soft/astrosat/resp/uvit/"])

Extract XSPEC/Sherpa compatible source and background PHA spectral files from AstroSat/UVIT broadband image generated from CCDLAB processing pipeline.

This function extracts source and background counts using the DS9 region files and converts them into PHA spectral files. For details on grating orders and spectral responses, see Dewangan (2021).

...

Arguments

Required parameters

target::String: Name of the target available in the observed UVIT field.

imagefile::String: Name of the FUV or NUV image file in FITS format generated using CCDLAB.

ds9srcregfile::String: Name of the ds9 region file for source.

ds9bgdregfile::String: Name of the ds9 region file for background.

Optional parameters

satu_corr::Bool: Apply saturation correction for point sources only if true (default).

respdir::String: Name of the directory containing the response matrices in the local machine. The response files can be downloaded from the GitHub page.

Output

Source and background PHA files.

Some relevant information are also printed on the screen.

...

UVITTools.uvit_saturation_corr — Function

uvit_saturation_corr(count_rate[, frames_per_sec=28.717])

Correct UVIT source count rate for saturation effect.

This function is based on the algorithm provided in Tandon et al. (2017), and used by uvit_aphot.jl for circular extraction region only.

UVITTools.uvit_zp_uc — Function

uvit_zp_uc(channel, uvit_filter)

Obtain the magnitude zero point and unit conversation factor for FUV or NUV channel in any filter.

This function is based on the calibration information from Tandon et al. (2017) and (2020). The function is used by uvit_aphot.jl to count rates to flux densities and magnitudes.

...

Arguments

channel::String: UVIT channel either FUV or NUV.

uvit_filter::String: UVIT filter name. Different names can be used for the same filter e.g., "F1", "F148W", "CaF2-1".

...

Miscellaneous

UVITTools.xycen_from_ds9reg — Function

xycen_from_ds9reg(ds9regfile)

Extract center coordinates (x,y) from a DS9 circular region file.

UVITTools.write_uvit_grating_phafile — Function

write_uvit_grating_phafile(uvit_detector, uvit_grating, channels, counts, exptime_sec[, phafile="phafile.pha"])

Write OGIP compatible PHA spectral file from grating count spectra - channel vs counts.

This function is used by grating spectroscopy tools. ...

Arguments

uvit_detector::String: FUV or NUV.

uvit_grating::String: One of the UVIT gratings - FUV-Grating1, FUV-Grating2, NUV-Grating.

channels::Array{Int64}: An array of spectral channels.

counts::Array{Float64}: An array of counts corresponding to the spectral channel array.

exposure_time_sec::Float64: Exposure time (s).

Optional

phafile::String: Name of the PHA spectral file to be generated.

...

UVITTools.lambdaA2keV — Function

Convert wavelength in Angstrom to Energy in keV.

Convert wavelength in Angstrom to Energy in keV.

UVITTools.lambdaA2ergs — Function

Convert wavelength in Å to eargs.

UVITTools.flux_density_photons2cgs — Function

Convert photon number density nλ to flux density fλ.

UVITTools.flux_density_cgs2photons — Function

Convert flux density fλ (ergs/cm2/s/Å) to photon number density nλ (photons/cm2/s/Å).